Extract ClanService request handling module
